requirements.txt
streamlit
tensorflow
flask
pillow
numpy
requests
flask-cors
